feat: require java 22 everywhere
