+++ build-extensions/src/main/kotlin/updater-extensions.kt
@@ -127,7 +127,7 @@ object UpdaterFiles {
   val moduleTargetDirectory: File = targetDirectory.resolve("modules")
 
   val nodeFile: File = gradleOutputJar("node", "cloudnet")
-  val launcherFile: File = gradleOutputJar("launcher/java17", "launcher")
+  val launcherFile: File = gradleOutputJar("launcher/java22", "launcher")
   val launcherPatcherFile: File = gradleOutputJar("launcher/patcher", "launcher-patcher")
 
   fun updaterOutputFile(filename: String, fileExtension: String = "jar", subdirectory: String? = null): File {
