--- launcher/java17/src/main/java/eu/cloudnetservice/launcher/java17/util/BootstrapUtil.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2019-2024 CloudNetService team & contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package eu.cloudnetservice.launcher.java17.util;
-
-import java.util.concurrent.TimeUnit;
-
-public final class BootstrapUtil {
-
-  private BootstrapUtil() {
-    throw new UnsupportedOperationException();
-  }
-
-  public static void waitAndExit() {
-    try {
-      TimeUnit.SECONDS.sleep(5);
-    } catch (InterruptedException ignored) {
-    }
-
-    // goodbye!
-    System.exit(1);
-  }
-}
