+++ launcher/java22/build.gradle.kts
