+++ launcher/java22/src/main/java/eu/cloudnetservice/launcher/java22/cnl/defaults/VarCnlCommand.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2019-2024 CloudNetService team & contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package eu.cloudnetservice.launcher.java22.cnl.defaults;
+
+import eu.cloudnetservice.launcher.java22.cnl.CnlCommand;
+import java.util.Queue;
+import lombok.NonNull;
+
+public final class VarCnlCommand implements CnlCommand {
+
+  @Override
+  public void execute(@NonNull Queue<String> arguments) {
+    System.setProperty(arguments.remove(), String.join(" ", arguments));
+    // remove all arguments
+    arguments.clear();
+  }
+}
