+++ launcher/java22/src/main/resources/launcher.cnl
