+++ launcher/java22/src/test/resources/test.cnl
