+++ modules/bridge/src/main/java/eu/cloudnetservice/modules/bridge/platform/fabric/FabricBridgeInitializer.java
@@ -32,7 +32,7 @@ import lombok.NonNull;
   dependencies = {
     @Dependency(name = "fabricloader", version = ">=0.15.0"),
     @Dependency(name = "minecraft", version = "~1.21"),
-    @Dependency(name = "java", version = ">=21")
+    @Dependency(name = "java", version = ">=22")
   },
   authors = "CloudNetService"
 )
