+++ modules/dockerized-services/src/main/java/eu/cloudnetservice/modules/docker/DockerizedServicesModule.java
@@ -49,7 +49,7 @@ public class DockerizedServicesModule extends DriverModule {
       () -> new DockerConfiguration(
         "docker-jvm",
         "host",
-        DockerImage.builder().repository("azul/zulu-openjdk").tag("17-jre-headless").build(),
+        DockerImage.builder().repository("azul/zulu-openjdk").tag("22-jre-headless").build(),
         Set.of(),
         Set.of(),
         Set.of(),
