+++ node/src/main/resources/lang/de_DE.properties
@@ -259,7 +259,7 @@ command-groups-rename-success=Die Gruppe {0$group$} wurde erfolgreich zu {1$new$
 #
 # Tasks
 #
-cloudnet-load-task-unsupported-java-version=Der Java Befehl wird von Task {0$task$} entfernt da er keine valide Java-Installation ist oder nicht kompatibel mit Java 17 ist
+cloudnet-load-task-unsupported-java-version=Der Java Befehl wird von Task {0$task$} entfernt da er keine valide Java-Installation ist oder nicht kompatibel mit Java 22 ist
 #
 # Command Tasks
 #
