+++ node/src/main/resources/lang/en_US.properties
@@ -257,7 +257,7 @@ command-groups-rename-success=The group {0$group$} was successfully renamed to {
 #
 # Tasks
 #
-cloudnet-load-task-unsupported-java-version=Removing java command from {0$task$} because it isn't a valid java installation or incompatible with Java 17
+cloudnet-load-task-unsupported-java-version=Removing java command from {0$task$} because it isn't a valid java installation or incompatible with Java 22
 #
 # Command Tasks
 #
